# openapi.yaml
openapi: 3.1.0
info:
  title: Quote Builder API
  description: Generate government fleet quotes as downloadable PDFs using vehicle and pricing data.
  version: 1.1.0
servers:
  - url: https://quote-api-qflm.onrender.com
paths:
  /api/generate-quote:
    post:
      operationId: generateQuote
      summary: Submit quote generation request and receive DocRaptor job ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - customer
                - customerContactName
                - customerPhone
                - customerEmail
                - dealership
                - managerName
                - managerPhone
                - managerEmail
                - vehicles
                - transport
              properties:
                quoteNumber:
                  type: string
                quoteDate:
                  type: string
                  format: date
                quoteExpires:
                  type: string
                  format: date
                customer:
                  type: string
                customerContactName:
                  type: string
                customerPhone:
                  type: string
                customerEmail:
                  type: string
                dealership:
                  type: string
                  enum:
                    - Sames Laredo Chevrolet
                    - Sames Bastrop Ford
                    - Sames Bastrop CDJ
                    - Sames McAllen Ford
                managerName:
                  type: string
                managerPhone:
                  type: string
                managerEmail:
                  type: string
                vehicles:
                  type: array
                  items:
                    type: object
                    required:
                      - year
                      - make
                      - model
                      - contract
                      - quantity
                      - msrp
                      - discountPrice
                      - taxAndLicense
                      - totalPrice
                      - color
                      - standardOptions
                    properties:
                      year:
                        type: string
                      make:
                        type: string
                      model:
                        type: string
                      contract:
                        type: string
                      quantity:
                        type: integer
                      msrp:
                        type: number
                      discountPrice:
                        type: number
                      taxAndLicense:
                        type: number
                      totalPrice:
                        type: number
                      color:
                        type: string
                      standardOptions:
                        type: array
                        items:
                          type: string
                upgrades:
                  type: array
                  items:
                    type: object
                    required:
                      - name
                      - quantity
                      - price
                      - total
                    properties:
                      name:
                        type: string
                      quantity:
                        type: integer
                      price:
                        type: number
                      total:
                        type: number
                upfitter:
                  type: object
                  required:
                    - company
                    - quoteNumber
                    - description
                    - total
                  properties:
                    company:
                      type: string
                    quoteNumber:
                      type: string
                    description:
                      type: string
                    total:
                      type: number
                transport:
                  type: object
                  required:
                    - miles
                    - ratePerMile
                    - total
                  properties:
                    miles:
                      type: integer
                    ratePerMile:
                      type: number
                    total:
                      type: number
                tradeIn:
                  type: object
                  properties:
                    vin:
                      type: string
                    allowance:
                      type: number
      responses:
        '200':
          description: Job submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  quoteNumber:
                    type: string
                  statusId:
                    type: string
                  filename:
                    type: string
        '500':
          description: Internal server error

  /api/quote-status/{statusId}:
    get:
      operationId: getQuoteStatus
      summary: Check the status of a quote PDF generation
      parameters:
        - name: statusId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: DocRaptor job status
          content:
            application/json:
              schema:
                type: object
                properties:
                  done:
                    type: boolean
                  download_url:
                    type: string
                    nullable: true
                  error:
                    type: string
                    nullable: true
        '500':
          description: Failed to retrieve job status